Italy/requirements.txt
python-dotenv
curl_cffi
urllib3
